Move simulation related packages to `sim`
Refactor the top-level packages to move all simulation related packages
to `sim`.

The changes here make no functional alteration; they only move packages.
