+++ README.md
@@ -2,7 +2,7 @@
 
 This repository contains a Go implementation of the [Filecoin finality module,
 F3](https://github.com/filecoin-project/FIPs/blob/f3564189d11817328168c9e75a80ff5f7292ba13/FIPS/fip-xxxx.md)
-(WIP). 
+(WIP).
 
 This executes an iterative GossiPBFT consensus protocol to declare tipsets as
 final when voted for by >2/3 of the consensus power.
@@ -27,24 +27,24 @@ There is also a main entry point to run GossiPBFT instance with honest nodes in
 simulation.
 
 ```
-$ go run f3sim.go -help
-Usage of /path/to/f3sim:
+$ go run ./sim/cmd/f3sim -help
+Usage of f3sim:
+  -delta-back-off-exponent float
+        exponential factor adjusting the delta value per round (default 1.3)
   -granite-delta float
-    	granite delta parameter (default 6)
-  -granite-delta-rate float
-    	change in delta for each round (default 2)
+        granite delta parameter (bound on message delay) (default 2)
   -iterations int
-    	number of simulation iterations (default 1)
+        number of simulation iterations (default 1)
   -latency-mean float
-    	mean network latency (default 0.5)
+        mean network latency in seconds (default 0.5)
   -latency-seed int
-    	random seed for network latency (default current time)
-  -max-rounds int
-    	max rounds to allow before failing (default 10)
+        random seed for network latency (default <current time>)
+  -max-rounds uint
+        max rounds to allow before failing (default 10)
   -participants int
-    	number of participants (default 3)
+        number of participants (default 3)
   -trace int
-    	trace verbosity level
+        trace verbosity level
 ```
 
 ## Integration
@@ -53,7 +53,9 @@ The code does not yet express an API for integration into a Filecoin node.
 Coming soon!
 
 ## Structure
+
 Modules:
+
 - `f3`: the protocol implementation
 - `sim`: the simulation harness
 - `adversary`: specific adversarial behaviors for use in tests
