+++ sim/adversary/absent.go
