+++ sim/adversary/decide.go
