+++ sim/adversary/withhold.go
