+++ test/absent_test.go
@@ -3,8 +3,8 @@ package test
 import (
 	"testing"
 
-	"github.com/filecoin-project/go-f3/adversary"
 	"github.com/filecoin-project/go-f3/sim"
+	"github.com/filecoin-project/go-f3/sim/adversary"
 	"github.com/stretchr/testify/require"
 )
